[ALPS04655831] mm: mm: fix possioble integer overflows
Coverty reports a possible integer overflows. Fix them
in this change.
Unintentional integer overflow: 8246721

MTK-Commit-Id: 7693aea0c30d33038d423e5d03f18bdc8043b72c

Feature: [Module]Kernel Memory Management
CR-Id: ALPS04655831

+++ mm/page_alloc.c
@@ -6611,7 +6611,8 @@ void __init mem_init_print_info(const char *str)
 #ifdef CONFIG_MTK_MEMCFG
 		kernel_reserve_meminfo.available =
 			(unsigned long long)(nr_free_pages()) << PAGE_SHIFT;
-		kernel_reserve_meminfo.total = physpages << PAGE_SHIFT;
+		kernel_reserve_meminfo.total =
+			(unsigned long long)physpages << PAGE_SHIFT;
 		kernel_reserve_meminfo.kernel_code = codesize;
 		kernel_reserve_meminfo.rwdata = datasize;
 		kernel_reserve_meminfo.rodata = rosize;
